@@ -18,18 +18,18 @@
   },
   "dependencies": {
     "@graasp/ui": "link:..",
-    "@testing-library/jest-dom": "link:../node_modules/@testing-library/jest-dom",
-    "@testing-library/react": "link:../node_modules/@testing-library/react",
-    "@testing-library/user-event": "link:../node_modules/@testing-library/user-event",
-    "@types/jest": "link:../node_modules/@types/jest",
-    "@types/node": "link:../node_modules/@types/node",
-    "@types/react": "link:../node_modules/@types/react",
-    "@types/react-dom": "link:../node_modules/@types/react-dom",
-    "immutable": "4.0.0",
-    "react": "17.0.2",
-    "react-dom": "17.0.2",
-    "react-scripts": "3.4.4",
-    "typescript": "4.6.4"
+    "@testing-library/jest-dom": "5.16.4",
+    "@testing-library/react": "13.3.0",
+    "@testing-library/user-event": "14.2.0",
+    "@types/jest": "28.1.0",
+    "@types/node": "17.0.39",
+    "@types/react": "18.0.10",
+    "@types/react-dom": "18.0.5",
+    "immutable": "4.1.0",
+    "react": "18.1.0",
+    "react-dom": "18.1.0",
+    "react-scripts": "5.0.1",
+    "typescript": "4.7.2"
   },
   "devDependencies": {
     "@babel/plugin-syntax-object-rest-spread": "^7.8.3",
