@@ -50,11 +50,11 @@
     "@graasp/utils": "github:graasp/graasp-utils.git",
     "clsx": "1.1.1",
     "http-status-codes": "2.2.0",
-    "immutable": "4.0.0",
-    "katex": "0.15.3",
+    "immutable": "4.1.0",
+    "katex": "0.15.6",
     "qs": "6.10.3",
     "react-cookie-consent": "7.4.1",
-    "react-i18next": "11.16.9",
+    "react-i18next": "11.17.0",
     "react-quill": "1.3.5",
     "react-text-mask": "5.4.3",
     "uuid": "8.3.2"
@@ -71,33 +71,34 @@
     "react-dom": "18.1.0"
   },
   "devDependencies": {
-    "@babel/plugin-proposal-optional-chaining": "7.16.7",
-    "@babel/preset-env": "7.17.10",
-    "@babel/preset-typescript": "7.16.7",
-    "@commitlint/cli": "17.0.0",
-    "@commitlint/config-conventional": "17.0.0",
-    "@mui/icons-material": "^5.8.0",
-    "@mui/lab": "^5.0.0-alpha.83",
-    "@mui/material": "^5.8.1",
+    "@babel/plugin-proposal-optional-chaining": "7.17.12",
+    "@babel/preset-env": "7.18.2",
+    "@babel/preset-typescript": "7.17.12",
+    "@commitlint/cli": "17.0.2",
+    "@commitlint/config-conventional": "17.0.2",
+    "@mui/icons-material": "^5.8.2",
+    "@mui/lab": "^5.0.0-alpha.84",
+    "@mui/material": "^5.8.2",
+    "@mui/styles": "^5.8.0",
     "@rollup/plugin-babel": "5.3.1",
     "@rollup/plugin-commonjs": "22.0.0",
     "@rollup/plugin-node-resolve": "13.3.0",
     "@rollup/plugin-typescript": "8.3.2",
     "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.2.0",
+    "@testing-library/react": "^13.3.0",
     "@testing-library/user-event": "^14.2.0",
     "@types/enzyme": "3.10.12",
     "@types/enzyme-adapter-react-16": "1.0.6",
-    "@types/jest": "27.5.1",
+    "@types/jest": "28.1.0",
     "@types/katex": "0.14.0",
-    "@types/node": "^17.0.33",
+    "@types/node": "^17.0.39",
     "@types/qs": "6.9.7",
-    "@types/react": "18.0.9",
+    "@types/react": "18.0.10",
     "@types/react-dom": "18.0.5",
     "@types/react-text-mask": "5.4.11",
     "@types/uuid": "8.3.4",
-    "@typescript-eslint/eslint-plugin": "^2.34.0",
-    "@typescript-eslint/parser": "^5.23.0",
+    "@typescript-eslint/eslint-plugin": "^5.27.0",
+    "@typescript-eslint/parser": "5.27.0",
     "ag-grid-community": "27.3.0",
     "ag-grid-react": "27.3.0",
     "babel-jest": "28.1.0",
@@ -105,17 +106,16 @@
     "enzyme": "3.11.0",
     "enzyme-adapter-react-16": "1.15.6",
     "enzyme-to-json": "3.6.2",
-    "eslint": "^6.8.0",
-    "eslint-config-prettier": "^6.15.0",
-    "eslint-config-standard": "^14.1.1",
-    "eslint-config-standard-react": "^9.2.0",
+    "eslint": "^8.16.0",
+    "eslint-config-prettier": "^8.5.0",
+    "eslint-config-standard": "^17.0.0",
+    "eslint-config-standard-react": "^11.0.1",
     "eslint-plugin-import": "^2.26.0",
     "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-prettier": "^3.4.1",
-    "eslint-plugin-promise": "^4.3.1",
-    "eslint-plugin-react": "^7.29.4",
+    "eslint-plugin-prettier": "^4.0.0",
+    "eslint-plugin-promise": "^6.0.0",
+    "eslint-plugin-react": "^7.30.0",
     "eslint-plugin-react-app": "6.2.2",
-    "eslint-plugin-standard": "^4.1.0",
     "husky": "^8.0.1",
     "jest": "28.1.0",
     "node-sass": "7.0.1",
@@ -123,18 +123,18 @@
     "pinst": "^3.0.0",
     "prettier": "^2.6.2",
     "react": "18.1.0",
-    "react-query": "3.39.0",
-    "react-test-renderer": "17.0.2",
+    "react-query": "3.39.1",
+    "react-test-renderer": "18.1.0",
     "rimraf": "3.0.2",
-    "rollup": "2.73.0",
+    "rollup": "2.75.5",
     "rollup-jest": "2.0.0",
     "rollup-plugin-copy": "3.4.0",
     "rollup-plugin-peer-deps-external": "2.2.4",
     "rollup-plugin-scss": "3.0.0",
     "standard-version": "9.5.0",
-    "ts-jest": "28.0.2",
-    "ts-node": "10.7.0",
-    "typescript": "4.6.4"
+    "ts-jest": "28.0.4",
+    "ts-node": "10.8.0",
+    "typescript": "4.7.2"
   },
   "files": [
     "dist"
